Support wildcards in package names (as portage does).
Fallback to Atom() if dep_expand() does fail with InvalidPackageAtom as
the latter doesn't support wildcards (portage does as of 2.1.9). This
way, it is now possible to mangle wildcard entries with flaggie.

This also fixes support for explicit 'null/foobar' atom spec ('null/' is
used by dep_expand() to denote that it is unable to expand the category).